pyproject.toml
[tool.poetry]
name = "bpp_plugins.import_dbf"
version = "0.1.0"
description = "Plugin do BPP realizujący import bazy danych w formatach DBF"
authors = ["Michał Pasternak <michal.dtz@gmail.com>"]
readme = "README.rst"
packages = [
         {include = "bpp_plugins/import_dbf", from="src"}
         ]

[tool.poetry.dependencies]
python = ">=3.10,<4"
bpp-iplweb = ">=202307.1105"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
